--- examples/fan/main/app_driver.cpp
@@ -1,166 +1,0 @@
-/*
-   This example code is in the Public Domain (or CC0 licensed, at your option.)
-
-   Unless required by applicable law or agreed to in writing, this
-   software is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
-   CONDITIONS OF ANY KIND, either express or implied.
-*/
-
-#include <esp_log.h>
-#include <stdlib.h>
-#include <string.h>
-
-#include <device.h>
-#include <esp_matter.h>
-#include <led_driver.h>
-
-#include <app_priv.h>
-
-using namespace chip::app::Clusters;
-using namespace esp_matter;
-
-static const char *TAG = "app_driver";
-extern uint16_t fan_endpoint_id;
-
-/* Do any conversions/remapping for the actual value here */
-static esp_err_t app_driver_light_set_power(led_driver_handle_t handle, esp_matter_attr_val_t *val)
-{
-    return driver_set_power_gpio(handle, val->val.b); //led_driver_set_power(handle, val->val.b);
-}
-
-static esp_err_t app_driver_light_set_brightness(led_driver_handle_t handle, esp_matter_attr_val_t *val)
-{
-   // int value = REMAP_TO_RANGE(val->val.u8, MATTER_BRIGHTNESS, STANDARD_BRIGHTNESS);
-    return led_driver_set_brightness(handle, 100);
-}
-
-static esp_err_t app_driver_light_set_hue(led_driver_handle_t handle, esp_matter_attr_val_t *val)
-{
-    int value = REMAP_TO_RANGE(val->val.u8, MATTER_HUE, STANDARD_HUE);
-    return led_driver_set_hue(handle, value);
-}
-
-static esp_err_t app_driver_light_set_saturation(led_driver_handle_t handle, esp_matter_attr_val_t *val)
-{
-    int value = REMAP_TO_RANGE(val->val.u8, MATTER_SATURATION, STANDARD_SATURATION);
-    return led_driver_set_saturation(handle, value);
-}
-
-static esp_err_t app_driver_light_set_temperature(led_driver_handle_t handle, esp_matter_attr_val_t *val)
-{
-    uint32_t value = REMAP_TO_RANGE_INVERSE(val->val.u16, STANDARD_TEMPERATURE_FACTOR);
-    return led_driver_set_temperature(handle, value);
-}
-
-static void app_driver_button_toggle_cb(void *arg)
-{
-    ESP_LOGI(TAG, "Toggle button pressed");
-    uint16_t endpoint_id = fan_endpoint_id;
-    uint32_t cluster_id = FanControl::Id;
-    uint32_t attribute_id = FanControl::Attributes::FanMode::Id;
-
-    node_t *node = node::get();
-    endpoint_t *endpoint = endpoint::get(node, endpoint_id);
-    cluster_t *cluster = cluster::get(endpoint, cluster_id);
-    attribute_t *attribute = attribute::get(cluster, attribute_id);
-
-    esp_matter_attr_val_t val = esp_matter_invalid(NULL);
-    attribute::get_val(attribute, &val);
-    val.val.b = !val.val.b;
-       ESP_LOGI(TAG, "Value >>>>> %d", val.val.b);
-
-    
-    attribute::update(endpoint_id, cluster_id, attribute_id, &val);
-}
-
-esp_err_t app_driver_attribute_update(app_driver_handle_t driver_handle, uint16_t endpoint_id, uint32_t cluster_id,
-                                      uint32_t attribute_id, esp_matter_attr_val_t *val)
-{
-    esp_err_t err = ESP_OK;
-    if (endpoint_id == fan_endpoint_id) {
-        led_driver_handle_t handle = (led_driver_handle_t)driver_handle;
-        if (cluster_id == FanControl::Id) {
-            if (attribute_id == FanControl::Attributes::FanMode::Id) {
-                err = app_driver_light_set_power(handle, val);
-            }
-        } else if (cluster_id == FanControl::Id) {
-            if (attribute_id == FanControl::Attributes::PercentSetting::Id) {
-                err = app_driver_light_set_brightness(handle, val);
-            }
-        } else if (cluster_id == FanControl::Id) {
-            if (attribute_id == ColorControl::Attributes::CurrentHue::Id) {
-                err = app_driver_light_set_hue(handle, val);
-            } else if (attribute_id == ColorControl::Attributes::CurrentSaturation::Id) {
-                err = app_driver_light_set_saturation(handle, val);
-            } else if (attribute_id == ColorControl::Attributes::ColorTemperatureMireds::Id) {
-                err = app_driver_light_set_temperature(handle, val);
-            }
-        }
-    }
-    return err;
-}
-
-esp_err_t app_driver_light_set_defaults(uint16_t endpoint_id)
-{
-    esp_err_t err = ESP_OK;
-    void *priv_data = endpoint::get_priv_data(endpoint_id);
-    led_driver_handle_t handle = (led_driver_handle_t)priv_data;
-    node_t *node = node::get();
-    endpoint_t *endpoint = endpoint::get(node, endpoint_id);
-    cluster_t *cluster = NULL;
-    attribute_t *attribute = NULL;
-    esp_matter_attr_val_t val = esp_matter_invalid(NULL);
-
-    /* Setting brightness */
-    cluster = cluster::get(endpoint, FanControl::Id);
-    attribute = attribute::get(cluster, FanControl::Attributes::PercentSetting::Id);
-    attribute::get_val(attribute, &val);
-    err |= app_driver_light_set_brightness(handle, &val);
-
-    /* Setting color */
-    cluster = cluster::get(endpoint, FanControl::Id);
-    attribute = attribute::get(cluster, ColorControl::Attributes::ColorMode::Id);
-    attribute::get_val(attribute, &val);
-    if (val.val.u8 == EMBER_ZCL_COLOR_MODE_CURRENT_HUE_AND_CURRENT_SATURATION) {
-        /* Setting hue */
-        attribute = attribute::get(cluster, ColorControl::Attributes::CurrentHue::Id);
-        attribute::get_val(attribute, &val);
-        err |= app_driver_light_set_hue(handle, &val);
-        /* Setting saturation */
-        attribute = attribute::get(cluster, ColorControl::Attributes::CurrentSaturation::Id);
-        attribute::get_val(attribute, &val);
-        err |= app_driver_light_set_saturation(handle, &val);
-    } else if (val.val.u8 == EMBER_ZCL_COLOR_MODE_COLOR_TEMPERATURE) {
-        /* Setting temperature */
-        attribute = attribute::get(cluster, ColorControl::Attributes::ColorTemperatureMireds::Id);
-        attribute::get_val(attribute, &val);
-        err |= app_driver_light_set_temperature(handle, &val);
-    } else {
-        ESP_LOGE(TAG, "Color mode not supported");
-    }
-
-    /* Setting power */
-    cluster = cluster::get(endpoint, FanControl::Id);
-    attribute = attribute::get(cluster, FanControl::Attributes::PercentSetting::Id);
-    attribute::get_val(attribute, &val);
-    err |= app_driver_light_set_power(handle, &val);
-
-    return err;
-}
-
-app_driver_handle_t app_driver_light_init()
-{
-    /* Initialize led */
-    led_driver_config_t config = led_driver_get_config();
-    led_driver_handle_t handle = led_driver_init(&config);
-    return (app_driver_handle_t)handle;
-}
-
-app_driver_handle_t app_driver_button_init()
-{
-    /* Initialize button */
-    button_config_t config = button_driver_get_config();
-    button_handle_t handle = iot_button_create(&config);
-    iot_button_register_cb(handle, BUTTON_PRESS_DOWN, app_driver_button_toggle_cb);
-    return (app_driver_handle_t)handle;
-}
