+++ examples/switch/README.md
