+++ examples/switch/main/CMakeLists.txt
