+++ examples/switch/main/app_driver.cpp
@@ -1,0 +1,101 @@
+/*
+   This example code is in the Public Domain (or CC0 licensed, at your option.)
+
+   Unless required by applicable law or agreed to in writing, this
+   software is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+   CONDITIONS OF ANY KIND, either express or implied.
+*/
+
+#include <esp_log.h>
+#include <stdlib.h>
+#include <string.h>
+
+#include <device.h>
+#include <esp_matter.h>
+#include <led_driver.h>
+
+#include <app_priv.h>
+
+using namespace chip::app::Clusters;
+using namespace esp_matter;
+
+static const char *TAG = "app_driver";
+extern uint16_t switch_endpoint_id;
+
+/* Do any conversions/remapping for the actual value here */
+static esp_err_t app_driver_light_set_power(led_driver_handle_t handle, esp_matter_attr_val_t *val)
+{
+    return driver_set_power_gpio(handle, val->val.b); //led_driver_set_power(handle, val->val.b);
+}
+
+
+static void app_driver_button_toggle_cb(void *arg)
+{
+    ESP_LOGI(TAG, "Toggle button pressed");
+    uint16_t endpoint_id = switch_endpoint_id;
+    uint32_t cluster_id = OnOff::Id;
+    uint32_t attribute_id = OnOff::Attributes::OnOff::Id;
+
+    node_t *node = node::get();
+    endpoint_t *endpoint = endpoint::get(node, endpoint_id);
+    cluster_t *cluster = cluster::get(endpoint, cluster_id);
+    attribute_t *attribute = attribute::get(cluster, attribute_id);
+
+    esp_matter_attr_val_t val = esp_matter_invalid(NULL);
+    attribute::get_val(attribute, &val);
+    val.val.b = !val.val.b;
+    attribute::update(endpoint_id, cluster_id, attribute_id, &val);
+}
+
+esp_err_t app_driver_attribute_update(app_driver_handle_t driver_handle, uint16_t endpoint_id, uint32_t cluster_id,
+                                      uint32_t attribute_id, esp_matter_attr_val_t *val)
+{
+    esp_err_t err = ESP_OK;
+    if (endpoint_id == switch_endpoint_id) {
+        led_driver_handle_t handle = (led_driver_handle_t)driver_handle;
+        if (cluster_id == OnOff::Id) {
+            if (attribute_id == OnOff::Attributes::OnOff::Id) {
+                err = app_driver_light_set_power(handle, val);
+            } 
+        }
+    }
+    return err;
+}
+
+esp_err_t app_driver_light_set_defaults(uint16_t endpoint_id)
+{
+    esp_err_t err = ESP_OK;
+    void *priv_data = endpoint::get_priv_data(endpoint_id);
+    led_driver_handle_t handle = (led_driver_handle_t)priv_data;
+    node_t *node = node::get();
+    endpoint_t *endpoint = endpoint::get(node, endpoint_id);
+    cluster_t *cluster = NULL;
+    attribute_t *attribute = NULL;
+    esp_matter_attr_val_t val = esp_matter_invalid(NULL);
+
+    /* Setting power */
+    cluster = cluster::get(endpoint, OnOff::Id);
+    attribute = attribute::get(cluster, OnOff::Attributes::OnOff::Id);
+    attribute::get_val(attribute, &val);
+    err |= app_driver_light_set_power(handle, &val);
+
+
+    return err;
+}
+
+app_driver_handle_t app_driver_light_init()
+{
+    /* Initialize led */
+    led_driver_config_t config = led_driver_get_config();
+    led_driver_handle_t handle = led_driver_init(&config);
+    return (app_driver_handle_t)handle;
+}
+
+app_driver_handle_t app_driver_button_init()
+{
+    /* Initialize button */
+    button_config_t config = button_driver_get_config();
+    button_handle_t handle = iot_button_create(&config);
+    iot_button_register_cb(handle, BUTTON_PRESS_DOWN, app_driver_button_toggle_cb);
+    return (app_driver_handle_t)handle;
+}
