+++ examples/switch/main/app_priv.h
