+++ examples/switch/main/zap-generated/PluginApplicationCallbacks.h
