+++ examples/switch/main/zap-generated/empty_file.cpp
