+++ examples/switch/main/zap-generated/endpoint_config.h
