+++ examples/switch/partitions.csv
