+++ examples/switch/partitions_h2.csv
