+++ examples/switch/sdkconfig.defaults
