+++ examples/switch/sdkconfig.defaults.esp32h2
