+++ examples/switch/sdkconfig.defaults.esp32s2
