+++ examples/switch/sdkconfig.defaults.ext_plat_ci
